BasicAgnoAgent/requirements.txt
agno
yfinance
python-dotenv 
google-genai
duckduckgo-search